# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pepsi PM API
  version: 0.8.0
  description: |
    REST API — backend Node/Express ที่ `backend/`
    พอร์ตล็อก dev: **5000** (สอดภาคผนวก ค / INFRASTRUCTURE)
    Auth: `SKIP_AUTH=true` ใน dev ไม่ต้องส่ง Bearer; production ใช้ JWT HS256 + โหลดสิทธิ์จาก `pepsi_pm`
servers:
  - url: http://127.0.0.1:5000
    description: Local backend
...
paths:
...
  /api/v1/auth/dev-token:
    post:
      tags: [Auth]
      summary: Mint JWT (development only)
      description: |
        Disabled when `NODE_ENV=production`. Requires `DEV_AUTH_SECRET` and `JWT_SECRET` on server.
        Header `X-Dev-Auth-Secret` must match `DEV_AUTH_SECRET`. Body must reference an existing active `users.gpid` (e.g. migration `V002__demo_user.sql` → `demo`).
      parameters:
        - name: X-Dev-Auth-Secret
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gpid]
              properties:
                gpid:
                  type: string
                  example: demo
      responses:
        '200':
          description: Access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  tokenType:
                    type: string
                    example: Bearer
                  expiresIn:
                    type: string
                    example: 8h
        '401':
          description: Invalid dev secret
        '404':
          description: Unknown gpid or disabled in production
        '503':
          description: DEV_AUTH_SECRET / JWT_SECRET not configured
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user (JWT or SKIP_AUTH stub)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      gpid:
                        type: string
                      permissions:
                        type: array
                        items:
                          type: string
        '401':
          description: Unauthorized
  /api/v1/imports/{kind}:
    post:
      tags: [Imports]
      summary: Upload SAP export file (IW37N, Confirm WO, GI, GR)
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [iw37n, confirm_wo, gi, gr]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Accepted — `import_batches` row + rows inserted into `stg_*` staging tables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportAccepted'
        '400':
          description: Bad request (wrong kind, missing file, or unreadable spreadsheet)
        '401':
          description: Unauthorized
        '403':
          description: Missing import.run permission
        '413':
          description: File or row count over limit
        '409':
          description: Duplicate file SHA when `DEDUPE_REJECT_DUPLICATE_FILE_SHA=true`
  /api/v1/import-batches:
    get:
      tags: [Imports]
      summary: List import batches
      parameters:
        - name: source_kind
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImportBatch'
  /api/v1/import-batches/{id}/normalize/async:
    post:
      tags: [Imports]
      summary: Enqueue normalize job (worker process)
      description: |
        Inserts a row into `import_jobs`. Run `npm run worker` in another terminal to process.
        Same effect as synchronous `POST .../normalize` when the job completes.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnqueueResponse'
        '401':
          description: Unauthorized
        '403':
          description: Missing import.run permission
  /api/v1/import-batches/{id}/normalize:
    post:
      tags: [Imports]
      summary: Normalize staging rows for a batch into operational tables
      description: |
        Atomic transaction per batch. `iw37n` / `confirm_wo` upsert `work_orders` and ensure `equipments`;
        `confirm_wo` also writes `order_confirmations` (dedupe key `sap_line_key`, migration V003).
        `gi` / `gr` delete existing `goods_movements` for this `import_batch_id` then insert from `stg_mb51_row`,
        ensuring `materials`. Requires `import.run` when `SKIP_AUTH=false`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Normalize summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NormalizeBatchResult'
        '400':
          description: Invalid id, empty batch, or unsupported source_kind
        '401':
          description: Unauthorized
        '403':
          description: Missing import.run permission
        '404':
          description: Batch not found
  /api/v1/import-batches/{id}/errors:
    get:
      tags: [Imports]
      summary: Errors for a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Error rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImportError'
  /api/v1/jobs/normalize-batch:
    post:
      tags: [Jobs]
      summary: Enqueue normalize_batch job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [batchId]
              properties:
                batchId:
                  type: integer
                  minimum: 1
      responses:
        '202':
          description: Enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnqueueResponse'
        '403':
          description: Missing import.run
  /api/v1/jobs/kpi-snapshot:
    post:
      tags: [Jobs]
      summary: Enqueue kpi_snapshot job (materialize `kpi_daily_snapshots`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [snapshotDate]
              properties:
                snapshotDate:
                  type: string
                  format: date
                  example: '2026-05-04'
                plant:
                  type: string
                  default: ''
      responses:
        '202':
          description: Enqueued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEnqueueResponse'
        '403':
          description: Missing report.dashboard
  /api/v1/jobs/{id}:
    get:
      tags: [Jobs]
      summary: Job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job row
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJobStatus'
        '404':
          description: Not found
  /api/v1/dashboard/stats:
    get:
      tags: [Dashboard]
      summary: สรุปข้อมูลแดชบอร์ด (กราฟ)
      description: |
        รวมจำนวน work orders ตาม `system_status`, จำนวน import รายวัน 14 วันล่าสุด,
        และจำนวน import ตาม `source_kind` ใน 30 วันล่าสุด — ต้องล็อกอิน (`requireAuth`)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          description: Unauthorized
  /api/v1/dashboard/kpi-snapshots:
    get:
      tags: [Dashboard]
      summary: รายการ kpi_daily_snapshots (สำหรับกราฟ KPI)
      description: |
        คืน `metrics_json` ที่ materialize จาก job `kpi_snapshot` (worker) — เรียง `snapshot_date` ใหม่สุดก่อน
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 60
            maximum: 200
        - name: plant
          in: query
          description: กรอง plant (ว่างได้) — ถ้าไม่ส่ง ได้ทุก plant
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/KpiSnapshotItem'
                  requestId:
                    type: string
        '401':
          description: Unauthorized
...
components:
  schemas:
    JobEnqueueResponse:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          type: string
        status:
          type: string
          example: pending
        batchId:
          type: string
          description: Present for normalize/async from import-batches only
        message:
          type: string
        requestId:
          type: string
    ImportJobStatus:
      type: object
      properties:
        id:
          type: string
        jobType:
          type: string
        status:
          type: string
          enum: [pending, running, done, failed, cancelled]
        payload:
          type: object
        attempts:
          type: integer
        maxAttempts:
          type: integer
        lastError:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        requestId:
          type: string
    ImportAccepted:
      type: object
      properties:
        batchId:
          type: string
        sourceKind:
          type: string
        status:
          type: string
          enum: [success, failed]
        rowCountAccepted:
          type: integer
        rowCountRejected:
          type: integer
          description: Blank lines skipped (not written to staging)
        message:
          type: string
        requestId:
          type: string
    NormalizeBatchResult:
      type: object
      properties:
        batchId:
          type: string
        sourceKind:
          type: string
          example: iw37n
        workOrdersUpserted:
          type: integer
          description: Staging rows applied (upsert by order_number); duplicates in file count multiple times
        goodsMovementsInserted:
          type: integer
        rowsSkipped:
          type: integer
          description: Rows without required keys (e.g. missing order_no or material_no)
        orderConfirmationsInserted:
          type: integer
          description: From `confirm_wo` only — rows inserted into `order_confirmations`
        orderConfirmationRowsSkipped:
          type: integer
          description: Staging rows without resolvable `work_orders` row
        requestId:
          type: string
    ImportBatch:
      type: object
      properties:
        id:
          type: string
        source_kind:
          type: string
        source_file_name:
          type: string
        status:
          type: string
        started_at:
          type: string
          format: date-time
    ImportError:
      type: object
      properties:
        id:
          type: string
        source_row_number:
          type: integer
        error_code:
          type: string
        error_message:
          type: string
...
    KpiSnapshotItem:
      type: object
      properties:
        id:
          type: string
        snapshotDate:
          type: string
          format: date
        plant:
          type: string
        metrics:
          type: object
          description: โครงจาก worker — workOrderTotal, workOrdersBySystemStatus, goodsMovementsByKind, …
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    DashboardStats:
      type: object
      properties:
        workOrders:
          type: object
          properties:
            total:
              type: integer
            bySystemStatus:
              type: array
              items:
                type: object
                properties:
                  status:
                    type: string
                  count:
                    type: integer
        importsByDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              count:
                type: integer
        importsByKind:
          type: array
          items:
            type: object
            properties:
              sourceKind:
                type: string
              count:
                type: integer
        requestId:
          type: string